This PR adds E2E tests to verify scaling up for a vLLM-emulator deployment under generated load, and subsequent scale down once the load generation is stopped.

Focusing on the added tests:
```sh
------------------------------
Test vllme deployment with VariantAutoscaling should scale up optimized replicas when load increases
/Users/tom/Desktop/llm-d-communitydev/inferno-autoscaler_tom/test/e2e/e2e_test.go:513
  STEP: verifying initial state of VariantAutoscaling @ 08/11/25 10:56:28.415
  STEP: getting the service endpoint for load generation @ 08/11/25 10:56:28.418
  STEP: setting up port-forward to the vllme service @ 08/11/25 10:56:28.42
  STEP: waiting for port-forward to be ready @ 08/11/25 10:56:28.422
  STEP: starting load generation to create traffic @ 08/11/25 10:56:29.442
  STEP: waiting for load to be processed and scaling decision to be made @ 08/11/25 10:56:29.446
  STEP: verifying that the controller has updated the status @ 08/11/25 10:58:29.464
Load Profile - Arrival Rate: 50.67, Avg Length: 228.00
Current Allocation - Replicas: 2, Accelerator: A100, 
Desired Optimized Allocation - Replicas: 2, Accelerator: A100
Current replicas for Deployment - vllme-deployment: 2
• [121.056 seconds]
------------------------------
Test vllme deployment with VariantAutoscaling should scale down as load stops
/Users/tom/Desktop/llm-d-communitydev/inferno-autoscaler_tom/test/e2e/e2e_test.go:607
  STEP: stopping the load generation script @ 08/11/25 10:58:29.47
  STEP: waiting for scaling down decision to be made @ 08/11/25 10:58:29.47
  STEP: verifying that the controller has updated the status @ 08/11/25 10:59:29.481
Load Profile - Arrival Rate: 2.67, Avg Length: 228.00
Current Allocation - Replicas: 1, Accelerator: A100, 
Desired Optimized Allocation - Replicas: 1, Accelerator: A100
Current replicas for Deployment - vllme-deployment: 1
• [60.016 seconds]
```

Results:
```sh
------------------------------

Ran 12 of 12 Specs in 327.831 seconds
SUCCESS! -- 12 Passed | 0 Failed | 0 Pending | 0 Skipped
--- PASS: TestE2E (327.83s)
PASS
ok      github.com/llm-d-incubation/inferno-autoscaler/test/e2e 328.691s
```